+++ waggle-dance-api/src/main/java/com/hotels/bdp/waggledance/api/model/AbstractMetaStore.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2017 Expedia Inc.
+ * Copyright (C) 2016-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,10 +15,12 @@
  */
 package com.hotels.bdp.waggledance.api.model;
 
-import java.beans.Transient;
-
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
+
+import java.beans.Transient;
+import java.util.Collections;
+import java.util.List;
 
 import org.hibernate.validator.constraints.NotBlank;
 
@@ -35,6 +37,7 @@ import com.google.common.base.Objects;
 public abstract class AbstractMetaStore {
 
   private String databasePrefix;
+  private List<String> writableDatabaseWhitelist;
   private @NotBlank String name;
   private @NotBlank String remoteMetaStoreUris;
   private @Valid MetastoreTunnel metastoreTunnel;
@@ -47,6 +50,17 @@ public abstract class AbstractMetaStore {
     this.name = name;
     this.remoteMetaStoreUris = remoteMetaStoreUris;
     this.accessControlType = accessControlType;
+  }
+
+  public AbstractMetaStore(
+      String name,
+      String remoteMetaStoreUris,
+      AccessControlType accessControlType,
+      List<String> writableDatabaseWhitelist) {
+    this.name = name;
+    this.remoteMetaStoreUris = remoteMetaStoreUris;
+    this.accessControlType = accessControlType;
+    this.writableDatabaseWhitelist = writableDatabaseWhitelist;
   }
 
   public static FederatedMetaStore newFederatedInstance(String name, String remoteMetaStoreUris) {
@@ -106,6 +120,17 @@ public abstract class AbstractMetaStore {
     this.accessControlType = accessControlType;
   }
 
+  public List<String> getWritableDatabaseWhiteList() {
+    if (writableDatabaseWhitelist == null) {
+      return Collections.emptyList();
+    }
+    return Collections.unmodifiableList(writableDatabaseWhitelist);
+  }
+
+  public void setWritableDatabaseWhiteList(List<String> writableDatabaseWhitelist) {
+    this.writableDatabaseWhitelist = writableDatabaseWhitelist;
+  }
+
   @Transient
   public MetaStoreStatus getStatus() {
     return status;
@@ -143,6 +168,7 @@ public abstract class AbstractMetaStore {
         .add("remoteMetaStoreUris", remoteMetaStoreUris)
         .add("metastoreTunnel", metastoreTunnel)
         .add("accessControlType", accessControlType)
+        .add("writableDatabaseWhiteList", writableDatabaseWhitelist)
         .add("status", status)
         .toString();
   }